move metadata to its own folder
