+++ node-api/src/metadata/mod.rs
@@ -1,0 +1,23 @@
+/*
+	Copyright 2021 Supercomputing Systems AG
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+		http://www.apache.org/licenses/LICENSE-2.0
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+*/
+
+mod error;
+mod from_v14_to_v15;
+mod metadata_types;
+
+pub use error::*;
+pub use from_v14_to_v15::v14_to_v15;
+pub use metadata_types::*;
+
+#[cfg(feature = "std")]
+mod print_metadata;
