+++ node-api/src/metadata/print_metadata.rs
